hide mobile menu
Added detect if on mobile then hide menu and hide again when scrolled
back to absolute top of page.
Detect menu height and add as padding to first section on page to
restore size *(height)* in case the first  section was fullpage and now
didn't cover the whole viewport in mobile anymore.